requirements.txt
Django>=3.2
gunicorn
whitenoise
google-generativeai
python-pptx

